create initial client, refactor Makefile, add csrf protection optional flag

+++ internal/pkg/httpsrv/handler_websocket.go
@@ -13,15 +13,17 @@ import (
 )
 
 func (s *Server) handlerWebSocket(w http.ResponseWriter, r *http.Request) {
-	// Get csrfToken
-	csrfToken := mux.Vars(r)["csrfToken"]
+	if s.csrfProtection {
+		// Get csrfToken
+		csrfToken := mux.Vars(r)["csrfToken"]
 
-	if !s.isValidCsrfToken(csrfToken) {
-		s.error(w, http.StatusInternalServerError, fmt.Errorf("bad csrf token"))
-		return
+		if !s.isValidCsrfToken(csrfToken) {
+			s.error(w, http.StatusInternalServerError, fmt.Errorf("bad csrf token"))
+			return
+		}
+		// Delete csrfToken now in order to prevent new connections
+		s.deleteCsrfToken(csrfToken)
 	}
-	// Delete csrfToken now in order to prevent new connections
-	s.deleteCsrfToken(csrfToken)
 
 	// Create and start a watcher.
 	var watch = watcher.New()
